Extract topic slot predicate matching
